feat: finished Buy Confirm Dialog
