+++ app/src/main/java/org/feature/fox/coffee_counter/ui/items/AddItemDialog.kt
@@ -1,6 +1,10 @@
 package org.feature.fox.coffee_counter.ui.items
 
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
