+++ app/src/main/res/values/strings.xml
@@ -45,4 +45,5 @@
     <string name="not_enough_funding">You don\'t have enough funding to add this to your cart</string>
     <string name="fill_all_fields">Please fill all fields</string>
     <string name="create_user">Create User</string>
+    <string name="buy">Buy</string>
 </resources>
